re-init cont.
