+++ limen-core/src/queue.rs
@@ -7,7 +7,8 @@ use crate::errors::QueueError;
 use crate::message::{payload::Payload, Message};
 use crate::policy::{AdmissionDecision, EdgePolicy, WatermarkState};
 
-pub mod descriptor;
+pub mod bench;
+pub mod link;
 
 pub mod spsc_array;
